[REFACTOR] Move source code to package, move resources
