+++ .gitignore
@@ -1,4 +1,4 @@
-/resources/dataset/dssp*
+.idea/
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
