+++ examples/resources/blocks/PDZA.blocks.txt
