+++ examples/resources/blocks/PDZB.blocks.txt
