+++ examples/resources/blocks/blocksToFasta.py
