+++ examples/resources/blosum/blosum100.iij
