+++ examples/resources/blosum/blosum30.iij
