+++ examples/resources/blosum/blosum35.iij
