+++ examples/resources/blosum/blosum40.iij
