+++ examples/resources/blosum/blosum45.iij
