+++ examples/resources/blosum/blosum50.iij
