+++ examples/resources/blosum/blosum55.iij
