+++ examples/resources/blosum/blosum60.iij
