+++ examples/resources/blosum/blosum62.iij
