+++ examples/resources/blosum/blosum65.iij
