+++ examples/resources/blosum/blosum70.iij
