+++ examples/resources/blosum/blosum80.iij
