+++ examples/resources/blosum/blosum85.iij
