+++ examples/resources/blosum/blosum95.iij
