+++ examples/resources/blosum/blosumn.iij
