+++ examples/resources/fasta/PDZ-A.fasta
