+++ examples/resources/fasta/PDZ-B.fasta
