+++ examples/resources/fasta/PDZ-sequences.fasta
