+++ examples/resources/fasta/SH3-A.fasta
