+++ examples/resources/fasta/SH3-B.fasta
