+++ examples/resources/fasta/SH3-C.fasta
