+++ examples/resources/fasta/SH3-D.fasta
