+++ examples/resources/fasta/SH3-sequence.fasta
