+++ examples/resources/fasta/example.fasta
