+++ examples/resources/fasta/maguk-sequences.fasta
