+++ examples/resources/fasta/msaresults-MUSCLE.fasta
