+++ examples/resources/fasta/msaresults-OMEGA.fasta
