+++ examples/resources/fasta/to-be-aligned.fasta
