+++ pyprot/protein.py
