--- resources/dataset/CATH_info_test-PARSED.txt
@@ -1,15 +1,0 @@
-> 1ARL|APO-CARBOXYPEPTIDASE A=ALPHA= (COX)|BOS TAURUS
-ARSTNTFNYATYHTLDEIYDFMDLLVAEHPQLVSKLQIGRSYEGRPIYVLKFSTGGSNRPAIWIDLGIHSREWITQATGVWFAKKFTEDYGQDPSFTAILDSMDIFLEIVTNPDGFAFTHSQNRLWRKTRSVTSSSLaVGVDANRNWDAGFGKAGASSSPaSETYHGKYANSEVEVKSIVDFVKDHGNFKAFLSIHSYSQLLLYPYGYTTQSIPDKTELNQVAKSAVAALKSLYGTSYKYGSIITTIYQASGGSIDWSYNQGIKYSFTFELRDTGRYGFLLPASQIIPTAQETWLGVLTIMEHTVNN
-CCCTTTCCTTCCCCHHHHHHHHHHHHHHCTTTEEEEEEEECTTCCEEEEEEECCCCCCCCEEEEEECCCTTCHHHHHHHHHHHHHHHHHTTTCHHHHHHHHHCEEEEECCCCHHHHHHHHHTCTTCCCCCCCCCCCCCCCCCHHHCCCCCTTCCCCECCTTCTTECCCCTTCCHHHHHHHHHHHHHCCEEEEEEEEECCCEEEECCCCCCCCCTTHHHHHHHHHHHHHHHHHHHCCCCEEEEHHHHCCCCCCCHHHHHHHTTCCEEEEEEECCCCCCHHHCCHHHHHHHHHHHHHHHHHHHHHHHHC
-> 1AVA|BARLEY ALPHA-AMYLASE 2(CV MENUET)|HORDEUM VULGARE
-ADPPPVHDTDGHELRADANYYVLSANRAHGGGLTMAPGHGRHaPLFVSQDPNGQHDGFPVRITPYGVAPSDKIIRLSTDVRISFRAYTTaLQSTEWHIDSELAAGRRHVITGPVKDPSPSGRENAFRIEKYSGAEVHEYKLMSbGDWbQDLGVFRDLKGGAWFLGATEPYHVVVFKKAPPA
-CCCCECECTTCCECETTCEEEEEECCHHHCCCEEEEEETTEEEEEEEEECCCTTCCCCCEEEEETTTCCCCCECETTCCEEEEECCCCTTCCCCECEECCCCETTECEEECCCCCCCCCCCHHHCEEEEECECCCCCCEEEEEECCCEEECEEECCCTTCCCEEECCCCECCEEEEEECCC
-> 1AVM|SUPEROXIDE DISMUTASE|PROPIONIBACTERIUM FREUDENREICHII SUBSP .
-AVYTLPELPYDYSALEPYISGEIMELHHDKHHKAYVDGANTALDKLAEARDKADFGAINKLEKDLAFNLAGHVNHSVFWKNMAPKGSAPERPTDELGAAIDEFFGSFDNMKAQFTAAATGIQGSGWASLVWDPLGKRINTLQFYDHQNNLPAGSIPLLQLDMWEHAFYLQYKNVKGDYVKSWWNVVNWDDVALRFSEARVA
-CCCCCCCCCCCTTTTTTTCCHHHHHHHHHTHHHHHHHHHHHHHHHHHHHHHHTCCTTHHHHHHHHHHHHHHHHHHHHHHHTECCTTCCCCCCCHHHHHHHHHHHCCHHHHHHHHHHHHHTCCCCEEEEEEEETTTTEEEEEEEETTTECCCTTCEEEEEEECCHHHTHHHHTTCHHHHHHHHHHHECHHHHHHHHHHHTCC
-> 1HGE|HEMAGGLUTININ, (G135R), HA1 CHAIN|INFLUENZA A VIRUS
-QDLPGNDNSTATLaLGHHAVPNGTLVKTITDDQIEVTNATELVQSSSTGKIbNNPHRILDGIDcTLIDALLGDPHcDVFQNETWDLFVERSKAFSNdYPYDVPDYASLRSLVASSGTLEFITEGFTWTGVTQNGRSNAdKRGPGSGFFSRLNWLTKSGSTYPVLNVTMPNNDNFDKLYIWGIHHPSTNQEQTSLYVQASGRVTVSTRRSQQTIIPNIGSRPWVRGLSSRISIYWTIVKPGDVLVINSNGNLIAPRGYFKMRTGKSSIMRSDAPIDTbISEeITPNGSIPNDKPFQNVNKITYGAePKYVKQNTLKLATGMRNVPEKQT
-CCCCCCCCCCEEEEEEEECCCCCEEECCCCCCCEEECCEEECEECCCCCCEEEECCCEEECTTCCHHHHHHTCHHHHHHTTCECCEEEECTTCCCCCCCEECTTHHHHHHHHHHHCECCEEECCCCCTTEEEEECEEEEEETTEEECCTTEEEEEEETTECCCEEEEEECCCCCCEEEEEEEEECCCHHHHHHHHCCCCCEEEEECCCCEEEECCCCCCCCCETTECCEEEEEEEEECTTCEEEEEEEECEEEECEEEECCCCCCEEEECCCCEEEEECCEEETTEEECTTCCEECCCCCCEEECCEECCCCCCEEECCCECCCCCCC
-> 1HMO|HEMERYTHRIN|THEMISTE DYSCRITUM
-GFPIPDPYaWDISFRTFYTIIDDEHKTLFNGILLLSQADNADHLNELRRCTGKHFLNEQQLMQSSQYAGYAEHKKAHDDFIHKLDTWDGDVTYAKNWLVNHIKTIDFKYRGKI
-CCCCCCCCCCCHHHCCCCHHHHHHHHHHHHHHHHHHHCCCHHHHHHHHHHHHHHHHHHHHHHHHTTCTTHHHHHHHHHHHHHHHHTCCCCHHHHHHHHHHHHHHTHHHHTTTC
