--- resources/dataset/CATH_info_test.txt
@@ -1,5 +1,0 @@
-1ARLA Alpha/beta
-1AVAC Beta
-1AVMA Alpha
-1HGEA Alpha/beta
-1HMOA Alpha
